@@ -1,7 +1,7 @@
 <template>
   <section v-if="educationData" class="bottom-pad bottom-border">
     <p class="section--heading">Education</p>
-    <EducationItem v-for="(data, index) in educationData" :key="index" :data="data" />
+    <education-item v-for="(data, index) in educationData" :key="index" :data="data" />
   </section>
 </template>
 <script>
@@ -12,9 +12,9 @@
     components: {
       EducationItem
     },
-    computed: {
-      educationData() {
-        return benoData && benoData.Education ? benoData.Education : null;
+    setup() {
+      return {
+        educationData: benoData.Education
       }
     }
   }
